Refactored EVTEinheit into unit app
