+++ web/veranstaltung/admin.py
@@ -5,6 +5,5 @@ from . import models
 
 admin.site.register(models.Veranstaltung)
 admin.site.register(models.Adresse)
-admin.site.register(models.EVTEinheit)
 admin.site.register(models.Qualifikation)
